added-language-files
using dictioanry from https://github.com/titoBouzout/Dictionaries which is a bit outdated compared to the one found here: https://extensions.openoffice.org/en/project/swedish-dictionaries-apache-openoffice

Was not able to get the one from OpenOffice to work, but someone who understands dictionaries could probably fix this in the future. The one included in this commit works, and might be good enough.